+++ portfolio/CHANGELOG.rst
@@ -3,7 +3,7 @@ ChangeLog of Portfolio App
 
 Portfolio App v0.6.0 (Upcoming)
 --------------------------------
-* DataCenter assets
+* DataCenter, DataCenterCampus assets
 * Export to CSV
 
 Portfolio App v0.5.0 (25-09-2024)